requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2024.1.0
openai>=1.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
